Auto stash before merge of "feature/TCIN-60" and "origin/feature/TCIN-60"
