+++ example.env
@@ -3,4 +3,7 @@ JWT_EXP_MIN=yourtime
 JDATABASE="your/path/to/json/db.json"
 
 LOCAL_DATABASE_PORT=0000
+LOCAL_DATABASE_DIR=""
+
 WEB_SOCKET_PORT=0000
+EXPIRATION_TIME= 0
