+++ src/manager/processEnv/envEnum.ts
@@ -1,0 +1,4 @@
+export enum nodeEnv {
+    DEVELOP = 'development',
+    PROD = 'production',
+}
