+++ src/routes/v1/localCache.routes.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import {
-    deleteLocalCache,
-    deleteCachedData,
+    // deleteLocalCache,
+    // deleteCachedData,
     getLocalCache,
     patchLocalCache,
     postLocalCache,
@@ -12,9 +12,9 @@ import config from 'config';
 
 const router = Router();
 
-router.get('/clearcache/folder/*', deleteCachedData);
-
-router.get('/clearcache/content/*', deleteCachedData);
+// router.get('/clearcache/folder/*', deleteCachedData);
+//
+// router.get('/clearcache/content/*', deleteCachedData);
 
 router.get('*', setHeader([...config.get<string[]>('localDatabase.customHeaders')]), getLocalCache);
 
@@ -24,6 +24,6 @@ router.put('*', putLocalCache);
 
 router.patch('*', patchLocalCache);
 
-router.delete('*', deleteLocalCache);
+// router.delete('*', deleteLocalCache);
 
 export default router;
