+++ src/types/model/bodyResponse.model.ts
@@ -1,0 +1,15 @@
+export class BodyResponseModel {
+    error: boolean;
+    cache: boolean;
+    message: string;
+    data: any;
+    filenameCached: string;
+
+    constructor() {
+        this.error = false;
+        this.cache = false;
+        this.message = '';
+        this.data = null;
+        this.filenameCached = '';
+    }
+}
